Fix epoch tracking, clean repo structure
Epoch tracking:
- Fix epoch incrementing on every resume (dynamics + tokenizer)
- Save batch_idx in checkpoint for mid-epoch resume
- load_checkpoint now returns 4-tuple: (epoch, global_step, loss, batch_idx)
- Skip batches on resume instead of starting next epoch

Repo structure cleanup:
- Move sbatch files to slurm/ directory
- Move project docs (DREAMERV4_AUDIT, PREEMPTION, project spec) to docs/
- Archive stale audit docs to docs/audits_legacy/
- Remove code_review/ (outdated, references old tau convention)
- Remove lazygit binary and tar.gz from repo
- Remove stale one-off scripts (debug_tau_conditioning, vram_test.sh, etc.)
- Remove empty source/ directory
- Add eval_results/, samples/, lazygit to .gitignore
- Top-level now contains only: CLAUDE.md, README.md, pyproject.toml, paper PDF

+++ scripts/pretokenize_frames.py
@@ -174,7 +174,7 @@ def process_batch(model, batch, device):
             latents = latents.permute(0, 3, 1, 2)   # (B, 32, 16, 16)
 
     # Save each latent as numpy (much smaller files than torch.save)
-    latents = latents.cpu().numpy().astype(np.float16)
+    latents = latents.float().cpu().numpy().astype(np.float16)
     for i, output_path in enumerate(output_paths):
         output_path = Path(output_path)
         output_path.parent.mkdir(parents=True, exist_ok=True)
